Split createGatewayRuntime to 3 parts:
1. core mesh plugin
2. yoga options builder
3. mesh plugins configurator
